Make things work again.
So, in short. The WB_IO2 controls the power of the 3v3_s connection. The
g_vl53l0x.begin/end/VL53L0X_Off methods control that pin as well. This
allows that vl53l0x lib to correctly power on/off the sensor.

However, if we're trying to do a GNSS fix at the same time, then we end
up with breaking the power on that module.

So, we really need to do GNSS fix and the distance measure in two steps.
We can't power on GNSS, do some distance measure and then hope that the
GNSS has found a fix.

The current code does that and remains working.

The GNSS still does 'hot-start' as there's a VDD connected to V_BCKP
which keeps it 'warm'.