Add GPS V2 enclosure.
After we discovered that the tension strength of the thread holes on the lid was very bad, i started to think of a design where there's no 'pull' on a long object.
The result is a flipped box, where the base has become a simple lid and the thread inserts sit in the box, which is the top size.
There's some great benefits to this as it turns out.

We also incorporated the new battery size which is much smaller.