Fix issue with confirmation code & receiving updates
because our confirmation code now takes longer, it means we don't get to LOOP() with the right data to handle the recieve data.

This 'quick fix' makes sure we run handlereceiveddata whenever we should. This is a bit hacky, but we can't fully refactor the code now to fix this. realistically, the lorahelper class is just not working for us, as it's too tightly coupled to main.